+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ## HEAD
 
 Features:
-- now, your data source can be a function! See the last section of the [Array, object and function data sources](http://www.handsontable.com/demo/datasources.html) page for examples (fixes [#471](https://github.com/warpech/jquery-handsontable/pull/471), [#435](https://github.com/warpech/jquery-handsontable/pull/435) and [#261](https://github.com/warpech/jquery-handsontable/pull/261))
+- now, your data source can be a function! See the last section of the [Array, object and function data sources](http://www.handsontable.com/demo/datasources.html) page for examples (fixes [#471](https://github.com/warpech/jquery-handsontable/pull/471), [#435](https://github.com/warpech/jquery-handsontable/pull/435), [#261](https://github.com/warpech/jquery-handsontable/issues/261))
 - also the column `data` property can be a function as well! Again, see the last section of the [data sources](http://www.handsontable.com/demo/datasources.html) page for examples
 - new methods: `countEmptyRows`, `countEmptyCols`, `isEmptyRow`, `isEmptyCol`. You can define your own functions for `isEmptyRow` and `isEmptyCol` (see [README.md](https://github.com/warpech/jquery-handsontable) for details)
 
 Bugfix:
 - replace reference to non-existent Exception with Error ([#494](https://github.com/warpech/jquery-handsontable/pull/494))
+- textarea copyPaste covering page elements ([#488](https://github.com/warpech/jquery-handsontable/issues/488), [#490](https://github.com/warpech/jquery-handsontable/issues/490))
 
 ## [0.8.13](https://github.com/warpech/jquery-handsontable/tree/v0.8.13) (Mar 12, 2013)